Code Rewritten + New visuals for interpretation
Rewrote the code for readability and to clear some errors + improve the labeling of graphs. Also added new visuals and graphs for interpretation. I looked at total play time, distribution of age, comparing subscription status and more. Not all of these are necessary to are question but I figured they could be useful